[feature] Finish simple harmonizer

+++ loudmon/Source/loudmon/oscilloscope.h
@@ -10,8 +10,8 @@ class OscilloscopeComponent :public juce::Component {
     addAndMakeVisible(plot_);
     addAndMakeVisible(x_slider_);
     x_slider_.onValueChange = std::bind(&OscilloscopeComponent::on_slider_change, this);
-    x_slider_.setRange(1, max_samples);
-    x_slider_.setValue(max_samples);
+    x_slider_.setRange(1, static_cast<double>(max_samples));
+    x_slider_.setValue(static_cast<double>(max_samples));
     x_slider_.setSliderStyle(Slider::SliderStyle::LinearHorizontal);
 
     addAndMakeVisible(y_slider_);
